goodbye grunt & babel, hello webpack & es6
