+++ .gitignore
@@ -6,6 +6,7 @@ logs/
 log/
 results/
 .idea/
+dist/
 
 *.seed
 *.log
