--- .jshintrc
@@ -1,25 +1,0 @@
-{
-    "node": true,
-    "camelcase": true,
-    "curly": true,
-    "immed": true,
-    "latedef": true,
-    "newcap": true,
-    "noarg": true,
-    "undef": true,
-    "unused": true,
-    "strict": true,
-    "trailing": true,
-    "smarttabs": true,
-    "globals": {
-      "alert": false,
-      "colorConvert": false,
-      "Image": false,
-      "window": false,
-      "document": false,
-      "navigator": false,
-      "getUserMedia": false,
-      "$": false,
-      "_": false
-    }
-}
