--- Gruntfile.js
@@ -1,67 +1,0 @@
-module.exports = function(grunt) {
-
-  grunt.initConfig({
-
-    pkg: grunt.file.readJSON('package.json'),
-
-    concat: {
-      options: {
-        separator: "\n\n"
-      },
-      dist: {
-        src: [
-          'src/_intro.js',
-          'src/main.js',
-          'src/Options.js',
-          'src/Util.js',
-          'src/*.js',
-          'src/_outro.js'
-        ],
-        dest: 'dist/<%= pkg.name.replace(".js", "") %>.js'
-      }
-    },
-
-    uglify: {
-      options: {
-        banner: '/*! <%= pkg.name.replace(".js", "") %> <%= grunt.template.today("dd-mm-yyyy") %> */\n'
-      },
-      dist: {
-        files: {
-          'dist/<%= pkg.name.replace(".js", "") %>.min.js': ['<%= concat.dist.dest %>']
-        }
-      }
-    },
-
-    qunit: {
-      files: ['test/*.html']
-    },
-
-    jshint: {
-      files: ['dist/Halftone.js'],
-      options: {
-        globals: {
-          console: true,
-          module: true,
-          document: true
-        },
-        jshintrc: '.jshintrc'
-      }
-    },
-
-    watch: {
-      files: ['<%= jshint.files %>'],
-      tasks: ['concat', 'jshint', 'qunit']
-    }
-
-  });
-
-  grunt.loadNpmTasks('grunt-contrib-uglify');
-  grunt.loadNpmTasks('grunt-contrib-jshint');
-  grunt.loadNpmTasks('grunt-contrib-qunit');
-  grunt.loadNpmTasks('grunt-contrib-watch');
-  grunt.loadNpmTasks('grunt-contrib-concat');
-
-  grunt.registerTask('test', ['jshint', 'qunit']);
-  grunt.registerTask('default', ['concat', 'jshint', 'qunit', 'uglify']);
-
-};
