--- demo/assets/halftone.css
@@ -1,9 +1,0 @@
-canvas.renderer {
-  width: 100%;
-  height: 100%;
-}
-
-div#viewport{
-  width: 1280px;
-  height: 720px;
-}
