--- demo/assets/main.css
@@ -1,20 +1,0 @@
-/* Styles for Halftone demo
- *
- * Author: Brett Crowell
- */
-
-/* === resets === */
-
-/* === defaults === */
-
-/* === generic styles */
-
-/* === layout === */
-
-/* === modules === */
-
-/* === specific styles */
-
-/* === subsections === */
-
-/* === hacks / overrides === */
