--- demo/still-image-demo.html
@@ -1,67 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>Halftone demo</title>
-  <link rel="stylesheet" href="./assets/normalize.css" />
-  <link rel="stylesheet" href="./assets/main.css" />
-  <style>
-    /* begin demo embedded styles */
-
-    /* end demo embedded styles */
-  </style>
-</head>
-<body>
-<h1>Halftone demo</h1>
-
-<!-- begin demo markup -->
-
-<div id="viewport"></div>
-
-<!-- end demo markup -->
-
-<p>See the <a href="http://brettcrowell.github.io/Halftone">project homepage</a>.
-<p>Check out the <a href="https://github.com/brettcrowell/Halftone">project repo</a>.
-<p>Copyright 2014 Brett Crowell</p>
-
-<script src="../bower_components/getUserMedia/dist/getUserMedia.js"></script>
-<script src="../bower_components/color-convert/color-convert-0.5.1.min.js"></script>
-<script src="./../dist/Halftone.js"></script>
-
-<script>
-  /* begin demo script */
-
-  var cache = {
-
-    virtualDOM: [],
-    currentMatrix: [],
-    maxLumens: 0,
-    minLumens: 15
-
-  }
-
-  var source = new Halftone.ImageSource("./assets/demo_image.jpg");
-      encoder = new Halftone.RasterFrameEncoder(),
-      compressor = new Halftone.Compressor(),
-      renderer = new Halftone.CachedCanvasRenderer();
-
-  document.getElementById('viewport').appendChild(renderer.getElement());
-
-  var resolution = Halftone.Options.resolution;
-
-
-
-  setTimeout(function(){
-
-    var imageData = source.getFrame();
-
-    var currentFrame = encoder.encodeFrame(imageData, Halftone.Options.stagger);
-
-    var differenceMatrix = compressor.getDifferenceMatrix(undefined, currentFrame);
-
-    renderer.render(differenceMatrix)
-  }, 1000);
-
-  /* end demo script */
-</script>
-</body>
-</html>
