+++ package.json
@@ -1,24 +1,30 @@
 {
-  "name": "Halftone",
-  "version": "0.0.0",
-  "main": "dist/Halftone.min.js",
-  "description": "Javascript experiment which converts a live video stream into halftones",
+  "name": "halftone",
+  "version": "1.0.0",
+  "description": "Converts the image from your webcam to Halftones in real-time",
+  "main": "src/main.js",
+  "directories": {
+    "test": "test"
+  },
+  "scripts": {
+    "start": "./node_modules/.bin/webpack-dev-server",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
   "repository": {
     "type": "git",
-    "url": "git://github.com/brettcrowell/halftone.git"
+    "url": "git+https://github.com/brettcrowell/halftone.git"
   },
-  "keywords": [],
   "author": "Brett Crowell",
   "license": "MIT",
-  "readmeFilename": "README.md",
-  "dependencies": {},
+  "bugs": {
+    "url": "https://github.com/brettcrowell/halftone/issues"
+  },
+  "homepage": "https://github.com/brettcrowell/halftone#readme",
   "devDependencies": {
-    "grunt": "~0.4.1",
-    "grunt-contrib-concat": "~0.3.0",
-    "grunt-contrib-jshint": "~0.6.4",
-    "grunt-contrib-qunit": "~0.3.0",
-    "grunt-contrib-uglify": "~0.2.4",
-    "grunt-contrib-watch": "~0.5.3",
-    "grunt-modernizr": "^0.6.0"
+    "babel-core": "^6.17.0",
+    "babel-loader": "^6.2.5",
+    "babel-preset-es2015": "^6.16.0",
+    "webpack": "^1.13.2",
+    "webpack-dev-server": "^1.16.2"
   }
 }
