--- src/_intro.js
@@ -1,3 +1,0 @@
-(function(root, undefined) {
-
-  "use strict";
