--- src/_outro.js
@@ -1,1 +1,0 @@
-}(this));
